+++ WTC_Projects/Swingy/src/main/java/Models/Heros/HeroFactory.java
@@ -16,7 +16,7 @@ public class HeroFactory {
     public static Hero generateLoadedHero(String[] heroStats){
         Hero player = new Hero(heroStats[0], heroStats[1], Integer.parseInt(heroStats[4]), Integer.parseInt(heroStats[5]), Integer.parseInt(heroStats[6]));
         player.set_level(Integer.parseInt(heroStats[2]));
-        player.set_experience(Integer.parseInt(heroStats[3]));
+        player.set_experience(Double.parseDouble(heroStats[3]));
         return player;
     }
 }
